Resources basd authorization

+++ Restaurants.Application/Dishes/Conmmands/DeleteDishes/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Conmmands.DeleteDishes;
@@ -9,7 +11,8 @@ namespace Restaurants.Application.Dishes.Conmmands.DeleteDishes;
 public class DeleteDishesForRestaurantCommandHandler
     (ILogger<DeleteDishesForRestaurantCommand> logger,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository): IRequestHandler<DeleteDishesForRestaurantCommand>
+    IDishesRepository dishesRepository,
+     IRestaurantAuthorizationService authorizationService) : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
     public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -20,7 +23,12 @@ public class DeleteDishesForRestaurantCommandHandler
 
         var restaurant = await restaurantsRepository.GetAsync(request.RestaurantId)
             ?? throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
-            
+
+        if (!authorizationService.Authorize(restaurant, ResourceOperation.Update))
+        {
+            throw new ForbidException();
+        }
+
 
         await dishesRepository.DeleteRangeAsync(restaurant.Dishes);
